refactored URI conversion #2
