+++ src/main/java/com/exasol/bucketfs/url/UriConverter.java
@@ -2,11 +2,9 @@ package com.exasol.bucketfs.url;
 
 import java.net.URI;
 import java.net.URISyntaxException;
-import java.nio.file.Path;
 import java.util.regex.Pattern;
 
-import com.exasol.bucketfs.client.OsCheck;
-import com.exasol.bucketfs.client.OsCheck.OSType;
+import com.exasol.bucketfs.url.OsCheck.OSType;
 
 import picocli.CommandLine.ITypeConverter;
 
@@ -34,24 +32,6 @@ public class UriConverter implements ITypeConverter<URI> {
     @Override
     public URI convert(final String value) throws URISyntaxException {
         return new URI(normalize(value));
-    }
-
-    /**
-     * @param path path to return URI for
-     * @return {@link URI} instance
-     * @throws URISyntaxException if format of path does not conform to valid URI
-     */
-    public URI convert(final Path path) throws URISyntaxException {
-        return new URI(prefix(path) + suffix(path.toString()));
-    }
-
-    private String prefix(final Path path) {
-        final String slash = isWindows() && path.isAbsolute() ? "/" : "";
-        return FILE + slash;
-    }
-
-    private String suffix(final String path) {
-        return isWindows() ? path.replace('\\', '/') : path;
     }
 
     private String normalize(final String value) {
