+++ src/test/java/com/exasol/bucketfs/url/UriConverterTest.java
@@ -3,25 +3,22 @@ package com.exasol.bucketfs.url;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 
 import java.net.URISyntaxException;
-import java.nio.file.Path;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-import com.exasol.bucketfs.client.OsCheck.OSType;
+import com.exasol.bucketfs.url.OsCheck.OSType;
 
 class UriConverterTest {
 
     @ParameterizedTest
     @CsvSource(value = { //
-            "relative,  dir/file.txt,  file://dir/file.txt", //
-            "file only, file.txt,      file://file.txt", //
-            "absolute,  /dir/file.txt, file:///dir/file.txt", //
+            "relative,  dir/file.txt,  dir/file.txt", //
+            "file only, file.txt,      file.txt", //
+            "absolute,  /dir/file.txt, /dir/file.txt", //
     })
     void convert(final String testCase, final String input, final String expected) throws Exception {
         assertThat(convert(OSType.WINDOWS, input), equalTo(expected));
@@ -32,21 +29,14 @@ class UriConverterTest {
     void driveLetter() throws Exception {
         final String input = "c:/dir/file.txt";
         assertThat(convert(OSType.WINDOWS, input), equalTo("file:///" + input));
-        assertThat(convert(OSType.LINUX, input), equalTo("file://" + input));
+        assertThat(convert(OSType.LINUX, input), equalTo(input));
     }
 
     private String convert(final OSType osType, final String input) throws URISyntaxException {
-        final Path path = (osType == OSType.WINDOWS) //
-                ? pathMock(input.replace('/', '\\'), input.startsWith("c:"))
-                : pathMock(input, input.startsWith("/"));
+        final String path = (osType == OSType.WINDOWS) //
+                ? input.replace('/', '\\')
+                : input;
         return new UriConverter(osType).convert(path).toString();
-    }
-
-    private Path pathMock(final String input, final boolean isAbsolute) {
-        final Path result = mock(Path.class);
-        when(result.toString()).thenReturn(input);
-        when(result.isAbsolute()).thenReturn(isAbsolute);
-        return result;
     }
 
     @Test
